--- numpyro_forecast/models/__init__.py
@@ -1,6 +1,0 @@
-"""Validated example forecasting models."""
-
-from numpyro_forecast.models.hierarchical import HierarchicalForecaster
-from numpyro_forecast.models.univariate import UnivariateForecaster
-
-__all__ = ["HierarchicalForecaster", "UnivariateForecaster"]
