+++ pyproject.toml
@@ -50,6 +50,7 @@ dev = [
     "pytest-xdist",
     "nbmake>=1.5",
     "ipykernel>=6.29",
+    "graphviz>=0.20",
     "ruff>=0.4",
     "prek>=0.0.1",
     "ty>=0.0.1a1",
