+++ tests/test_examples.py
@@ -2,11 +2,11 @@
 
 import jax.numpy as jnp
 import pytest
+from example_models import HierarchicalForecaster, UnivariateForecaster
 from jax import Array, random
 
 from numpyro_forecast import Forecaster, eval_crps
 from numpyro_forecast.datasets import bart_available, load_bart_hierarchical, load_bart_weekly
-from numpyro_forecast.models import HierarchicalForecaster, UnivariateForecaster
 from numpyro_forecast.util import fourier_features
